connect track selection screen to current geopoint so nearby tracks can be selected
